% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ordered_tree.R
\name{Ordered_Tree}
\alias{Ordered_Tree}
\title{Isotonic Regression with tree-ordered isotonic constraints}
\usage{
Ordered_Tree(y, w)
}
\arguments{
\item{y}{regression vector}

\item{w}{weights (default is NULL)}
}
\value{
A list containing the following elements:
\item{\code{x_ordered}}{Optimal solution that satisfies the .}
}
\description{
Isotonic Regression with tree-ordered isotonic constraints
}
\examples{
library(IsotoneOptimization)
set.seed(10)
Ordered_Tree(rnorm(10), rep(1/10))

}
\references{
Wang, X., Ying, J., Cardoso, J. V. D. M., & Palomar, D. P. (2022, June).
Efficient Algorithms for General Isotone Optimization.
In Proceedings of the \emph{AAAI Conference on Artificial Intelligence}
(Vol. 36, No. 8, pp. 8575-8583).
}
\author{
Xiwen Wang, Daniel P. Palomar
}
