% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specific_isotonic_str.R
\name{Solve_isotone_binary_tree}
\alias{Solve_isotone_binary_tree}
\title{Isotonic Regression with binary-tree-ordered isotonic constraints}
\usage{
Solve_isotone_binary_tree(y, w = NULL)
}
\arguments{
\item{y}{data vector}

\item{w}{weights (default is NULL)}
}
\value{
A list containing the following elements:
\item{\code{x_ordered}}{Optimal solution that satisfies the .}
}
\description{
Isotonic Regression with binary-tree-ordered isotonic constraints
}
\examples{
library(IsotoneOptimization)
set.seed(10)
Solve_isotone_binary_tree(rnorm(10), rep(1,10))

}
\references{
Wang, X., Ying, J., Cardoso, J. V. M., & Palomar, D. P. (2022, June).
Efficient Algorithms for General Isotone Optimization.
In Proceedings of the \emph{AAAI Conference on Artificial Intelligence}
(Vol. 36, No. 8, pp. 8575-8583).
}
\author{
Xiwen Wang, Zé Vinícius, and Daniel P. Palomar
}
