% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAG_isotone_optimization.R
\name{Solve_isotone_DAG}
\alias{Solve_isotone_DAG}
\title{Isotonic Regression with isotonic constraints represented by arbitrary directed acyclic graph}
\usage{
Solve_isotone_DAG(y, w = rep(1, length(y)), Emat)
}
\arguments{
\item{y}{regression vector}

\item{w}{weights (default is NULL)}

\item{Emat}{An m*2 Matrix to define the edges. Each row, given as (i,j), represents an edge from i to j}
}
\value{
A list containing the following elements:
\item{\code{x_ordered}}{Optimal solution that satisfies the .}
}
\description{
Isotonic Regression with isotonic constraints represented by arbitrary directed acyclic graph
}
\examples{
library(IsotoneOptimization)
Emat <- t(matrix( c(1,2,  1,3,  1,4,  2,5,  2,6,  3,7,  4,8,  5,8,  6,9,  6,10),2,10))
Solve_isotone_DAG(rnorm(10), rep(1,10), Emat)

}
\references{
Wang, X., Ying, J., Cardoso, J. V. D. M., & Palomar, D. P. (2022, June).
Efficient Algorithms for General Isotone Optimization.
In Proceedings of the \emph{AAAI Conference on Artificial Intelligence}
(Vol. 36, No. 8, pp. 8575-8583).
}
\author{
Xiwen Wang, Zé Vinícius, and Daniel P. Palomar
}
